chore: rename app router handlers
