+++ lib/app-router/handlers/index.ts
@@ -1,1 +1,1 @@
-export * from './app';
+export * from './enable-draft';
